requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
ollama==0.6.0
orjson==3.11.3
pillow==11.3.0
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.2.1
pytz==2024.2
